+++ sdk/resourcemanager/msi/armmsi/constants.go
@@ -11,5 +11,5 @@ package armmsi
 
 const (
 	moduleName    = "armmsi"
-	moduleVersion = "v0.7.0"
+	moduleVersion = "v0.8.0"
 )