+++ sdk/resourcemanager/msi/armmsi/operations_client.go
@@ -50,7 +50,6 @@ func NewOperationsClient(credential azcore.TokenCredential, options *arm.ClientO
 }
 
 // NewListPager - Lists available operations for the Microsoft.ManagedIdentity provider
-// If the operation fails it returns an *azcore.ResponseError type.
 // Generated from API version 2022-01-31-preview
 // options - OperationsClientListOptions contains the optional parameters for the OperationsClient.List method.
 func (client *OperationsClient) NewListPager(options *OperationsClientListOptions) *runtime.Pager[OperationsClientListResponse] {